Update UserController.cs
Add validation

+++ WEB_API_Task_WebApplication/Controllers/UserController.cs
@@ -32,6 +32,17 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<User> Create(User item)
         {
+         if (item == null)
+            {
+                ModelState.AddModelError("", "No data for creating user");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+             
             _userRepository.Create(item);
 
             return _userRepository.Entities;
@@ -41,6 +52,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<User> Delete(int Id)
         {
+        
+          if (Id == null)
+            {
+                ModelState.AddModelError("", "No data for deleting user");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+        
             _userRepository.Delete(Id);
 
             return _userRepository.Entities;
@@ -50,6 +73,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<User> Update(User item) 
         {
+        
+            if (item == null)
+            {
+                ModelState.AddModelError("", "No data for updating user");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+              
             _userRepository.Update(item);
 
             return _userRepository.Entities;
